backend/pyproject.toml
[project]
name = "novel2gal-backend"
version = "0.1.0"
description = "Novel2Gal backend — novel parsing, multi-agent scene generation, asset orchestration"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "surrealdb>=0.4",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "Pillow>=11.0",
    "numpy>=2.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "browser-use>=0.12.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]
